+++ demisto_sdk/commands/common/tests/playbook_test.py
@@ -290,6 +290,48 @@ class TestPlaybookValidator:
         (PLAYBOOK_JSON_ID_NOT_EQUAL_TO_TASKID, False)
     ]
 
+    DEPRECATED_VALID = {"deprecated": True, "description": "Deprecated. Use the XXXX playbook instead."}
+    DEPRECATED_VALID2 = {"deprecated": True, "description": "Deprecated. Feodo Tracker no longer supports this feed "
+                                                            "No available replacement."}
+    DEPRECATED_VALID3 = {"deprecated": True, "description": "Deprecated. The playbook uses an unsupported scraping"
+                                                            " API. Use Proofpoint Protection Server v2 playbook"
+                                                            " instead."}
+
+    DEPRECATED_INVALID_DESC = {"deprecated": True, "description": "Deprecated."}
+    DEPRECATED_INVALID_DESC2 = {"deprecated": True, "description": "Use the ServiceNow playbook to manage..."}
+    DEPRECATED_INVALID_DESC3 = {"deprecated": True, "description": "Deprecated. The playbook uses an unsupported"
+                                                                   " scraping API."}
+    DEPRECATED_INPUTS = [
+        (DEPRECATED_VALID, True),
+        (DEPRECATED_VALID2, True),
+        (DEPRECATED_VALID3, True),
+        (DEPRECATED_INVALID_DESC, False),
+        (DEPRECATED_INVALID_DESC2, False),
+        (DEPRECATED_INVALID_DESC3, False)
+    ]
+
+    CONDITIONAL_SCRPT_WITH_DFLT_NXT_TASK = {"id": "Intezer - scan host", "version": -1,
+                                            "tasks":
+                                                {'1': {'type': 'condition',
+                                                       'scriptName': 'testScript',
+                                                       'nexttasks': {'#default#': []}}}}
+
+    CONDITIONAL_SCRPT_WITH_NO_DFLT_NXT_TASK = {"id": "Intezer - scan host", "version": -1,
+                                               "tasks":
+                                               {'1': {'type': 'condition',
+                                                      'scriptName': 'testScript',
+                                                      'nexttasks': {'1': []}}}}
+
+    CONDITION_TASK_WITH_ELSE = {'1': {'type': 'condition',
+                                      'scriptName': 'testScript',
+                                      'nexttasks': {'#default#': []}}}
+
+    CONDITION_TASK_WITHOUT_ELSE = {'1': {'type': 'condition',
+                                         'scriptName': 'testScript',
+                                                       'nexttasks': {'1': []}}}
+    IS_ELSE_IN_CONDITION_TASK = [(CONDITIONAL_SCRPT_WITH_NO_DFLT_NXT_TASK.get('tasks').get('1'), False),
+                                 (CONDITIONAL_SCRPT_WITH_DFLT_NXT_TASK.get('tasks').get('1'), True)]
+
     @pytest.mark.parametrize("playbook_json, id_set_json, expected_result", IS_SCRIPT_ID_VALID)
     def test_playbook_script_id(self, mocker, playbook, repo, playbook_json, id_set_json, expected_result):
         """
@@ -416,3 +458,97 @@ class TestPlaybookValidator:
         structure = mock_structure("", playbook_json)
         validator = PlaybookValidator(structure)
         validator._is_taskid_equals_id() is expected_result
+
+    @pytest.mark.parametrize("current, answer", DEPRECATED_INPUTS)
+    def test_is_valid_deprecated_playbook(self, current, answer):
+        """
+        Given
+            1. A deprecated playbook with a valid description according to 'deprecated regex' (including the replacement
+               playbook name).
+            2. A deprecated playbook with a valid description according to the 'deprecated no replacement regex'.
+            3. A deprecated playbook with a valid description according to 'deprecated regex' (including the replacement
+               playbook name, and the reason for deprecation.).
+            4. A deprecated playbook with an invalid description that isn't according to the 'deprecated regex'
+               (doesn't include a replacement playbook name, or declare there isn't a replacement).
+            5. A deprecated playbook with an invalid description that isn't according to the 'deprecated regex'
+               (doesn't start with the phrase: 'Deprecated.').
+            6. A deprecated playbook with an invalid description that isn't according to the 'deprecated regex'
+               (Includes the reason for deprecation, but doesn't include a replacement playbook name,
+               or declare there isn't a replacement).
+        When
+            - running is_valid_as_deprecated.
+
+        Then
+            - a playbook with an invalid description will be errored.
+        """
+        structure = mock_structure("", current)
+        validator = PlaybookValidator(structure)
+        validator.current_file = current
+        assert validator.is_valid_as_deprecated() is answer
+
+    @pytest.mark.parametrize("playbook_json, expected_result", [(CONDITIONAL_SCRPT_WITH_NO_DFLT_NXT_TASK, True)])
+    def test_verify_all_conditional_tasks_has_else_path(self, playbook_json, expected_result):
+        """
+        Given
+            - A playbook with a condition without a default task
+
+        When
+            - Running Validate playbook
+
+        Then
+            - Function returns true as this is an ignored error.
+        """
+        structure = mock_structure("", playbook_json)
+        validator = PlaybookValidator(structure)
+        assert validator.verify_condition_tasks_has_else_path() is expected_result
+
+    @pytest.mark.parametrize("playbook_task_json, expected_result", IS_ELSE_IN_CONDITION_TASK)
+    def test_verify_else_for_conditions_task(self, playbook_task_json, expected_result):
+        """
+        Given
+            - A playbook condition task with a default task
+            - A playbook condition task without a default task
+
+        When
+            - Running Validate playbook
+
+        Then
+            - Return True if the condition task has default path , else false
+        """
+        structure = mock_structure("", playbook_task_json)
+        validator = PlaybookValidator(structure)
+        assert validator._is_else_path_in_condition_task(task=playbook_task_json) is expected_result
+
+    def test_name_contains_the_type(self, pack):
+        """
+        Given
+            - An playbook with a name that contains the word "playbook".
+        When
+            - running name_not_contain_the_type.
+        Then
+            - Ensure the validate failed.
+        """
+
+        playbook = pack.create_playbook(yml={"name": "test_playbook"})
+
+        structure_validator = StructureValidator(playbook.yml.path)
+        validator = PlaybookValidator(structure_validator)
+
+        assert not validator.name_not_contain_the_type()
+
+    def test_name_does_not_contains_the_type(self, pack):
+        """
+        Given
+            - An playbook with a name that does not contains the "playbook" string.
+        When
+            - running name_not_contain_the_type.
+        Then
+            - Ensure the validate passes.
+        """
+
+        playbook = pack.create_playbook(yml={"name": "test"})
+
+        structure_validator = StructureValidator(playbook.yml.path)
+        validator = PlaybookValidator(structure_validator)
+
+        assert validator.name_not_contain_the_type()