+++ demisto_sdk/tests/integration_tests/download_integration_test.py
@@ -111,3 +111,25 @@ def test_integration_download_list_files(demisto_client):
     assert result.exit_code == 0
     assert not result.stderr
     assert not result.exception
+
+
+def test_integration_download_fail(demisto_client, tmp_path):
+    """
+    Given
+    - Script to download, that exists on the machine.
+    - Playbook to download, that doesn't exist on the machine.
+
+    When
+    - Running demisto-sdk download command.
+
+    Then
+    - Ensure that the exit code is 1, since the playbook was not downloaded.
+    """
+    env = Environment(tmp_path)
+    pack_path = join(DEMISTO_SDK_PATH, env.PACK_INSTANCE_PATH)
+    runner = CliRunner(mix_stderr=False)
+    result = runner.invoke(main, [DOWNLOAD_COMMAND, "-o", pack_path, "-i", "TestScript", "-i", "DummyPlaybook1", "-f"])
+    assert "FILE NAME       REASON" in result.output[350:1050]
+    assert "-----------  ---------------------------------------" in result.output
+    assert 'DummyPlaybook1  File does not exist in Demisto instance' in result.output
+    assert result.exit_code == 1